Fix save activity name test
